+++ payments-core/src/main/java/com/stripe/android/utils/FeatureFlags.kt
@@ -6,6 +6,7 @@ import com.stripe.android.BuildConfig
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 object FeatureFlags {
     val cardBrandChoice = FeatureFlag()
+    val customerSheetACHv2 = FeatureFlag()
 }
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
