+++ paymentsheet/src/test/java/com/stripe/android/customersheet/CustomerAdapterTest.kt
@@ -16,11 +16,14 @@ import com.stripe.android.paymentsheet.PrefsRepository
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.model.SavedSelection
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
+import com.stripe.android.testing.FeatureFlagTestRule
 import com.stripe.android.utils.FakeCustomerRepository
+import com.stripe.android.utils.FeatureFlags
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.advanceTimeBy
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.kotlin.any
@@ -38,9 +41,14 @@ class CustomerAdapterTest {
     private val application = ApplicationProvider.getApplicationContext<Application>()
     private val testDispatcher = UnconfinedTestDispatcher()
 
+    @get:Rule
+    val featureFlagTestRule = FeatureFlagTestRule(
+        featureFlag = FeatureFlags.customerSheetACHv2,
+        isEnabled = true,
+    )
+
     @Before
     fun setup() {
-        CustomerSheetACHV2Flag = true
         PaymentConfiguration.init(
             context = application,
             publishableKey = "pk_123",
@@ -144,7 +152,7 @@ class CustomerAdapterTest {
 
     @Test
     fun `When CustomerSheetACHV2Flag is disabled, retrievePaymentMethods does not request US Bank Account payment methods`() = runTest {
-        CustomerSheetACHV2Flag = false
+        featureFlagTestRule.setEnabled(false)
         val customerRepository = mock<CustomerRepository>()
 
         val adapter = createAdapter(
@@ -162,7 +170,6 @@ class CustomerAdapterTest {
 
     @Test
     fun `When CustomerSheetACHV2Flag is enabled, retrievePaymentMethods requests US Bank Account payment methods`() = runTest {
-        CustomerSheetACHV2Flag = true
         val customerRepository = mock<CustomerRepository>()
 
         val adapter = createAdapter(
